refactor: unify ili9486 and touch drivers
