+++ lib/sample_app/application.ex
@@ -39,8 +39,8 @@ defmodule SampleApp.Application do
         # Children for all targets except host
         # Starts a worker by calling: Target.Worker.start_link(arg)
         # {Target.Worker, arg},
-        SampleApp.ui_mod(),
-        SampleApp.touch_mod()
+        SampleApp.ui_child_spec(),
+        SampleApp.touch_child_spec()
       ]
     end
   end
